+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Dev
+* fix: bddgen env throws if there is no @cucumber/cucumber
+
 ## 7.1.0
 * fix: access to $tags fixture in non-bdd project, ([#189](https://github.com/vitalets/playwright-bdd/issues/189))
 * fix: make @retries:0 special tag work, ([#187](https://github.com/vitalets/playwright-bdd/issues/187))